GUACAMOLE-234: Reorganize configuration into its own package.
